cleaning data files on the next step in refactoring architecture
